fix las viewer

+++ gears/src/main/java/org/hortonmachine/gears/io/las/core/LasRecord.java
@@ -57,7 +57,7 @@ public class LasRecord {
     /**
      * Classification.
      */
-    public byte classification = -1;
+    public short classification = -1;
 
     /**
      * Gps timestamp
